requirements.txt
flask
python-dotenv
pyaudio
google-generativeai
deepgram-sdk